requirements.txt
# Core dependencies
bleak>=0.21.1
pycryptodome>=3.19.0

# Optional: JoyCon support (install with: pip install -r requirements.txt[joycon])
# joycon-python>=0.1.0
# hidapi>=0.14.0
# pyglm>=2.7.0

# Development dependencies (install with: pip install -r requirements.txt[dev])
# pytest>=8.3.4
# pytest-asyncio>=0.24.0
